move test to main crate
