+++ golem-worker-executor/src/lib.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod config;
 pub mod context;
 pub mod services;
